+++ src/components/Studio/StudioInfo.tsx
@@ -1,20 +1,13 @@
 /** @jsxImportSource @emotion/react */
 import Bookmark from '@components/Bookmark/Bookmark';
-import Button from '@components/Button/Button';
 import ShareButton from '@components/Share/ShareButton';
 import { css } from '@emotion/react';
 import { breakPoints } from '@styles/BreakPoint';
-import {
-  DividerStyle,
-  TypoBodyMdR,
-  TypoBodyMdSb,
-  TypoCapSmR,
-  TypoTitleMdSb,
-  TypoTitleXsM,
-} from '@styles/Common';
+import { DividerStyle, TypoBodyMdR, TypoBodyMdSb, TypoCapSmR, TypoTitleMdSb } from '@styles/Common';
 import variables from '@styles/Variables';
 import { useMediaQuery } from 'react-responsive';
 import { IStudioDetail } from 'types/types';
+import StudioOptions from './StudioOptions';
 
 interface IStudioInfo {
   id: string | undefined;
@@ -25,29 +18,14 @@ const StudioInfo = ({ data, id }: IStudioInfo) => {
   let today = new Date();
   const dayIndex = today.getDay() === 0 ? 6 : today.getDay() - 1;
   const isPc = useMediaQuery({ minWidth: breakPoints.pc });
-
-  const option = {
-    CHANGING_ROOM: '탈의실',
-    DRESSING_ROOM: '파우더룸',
-    HAIR_MAKEUP: '헤어, 메이크업 수정',
-    INDIVIDUAL_EDITING: '1:1 보정',
-    SUIT_RENTAL_FREE: '정장 대여',
-    ORIGINAL_FILES: '원본파일 제공',
-    PARKING_AREA: '주차',
-  };
-
-  const optionIcon = {
-    CHANGING_ROOM: '/img/icon-room.svg',
-    DRESSING_ROOM: '/img/icon-powder.svg',
-    HAIR_MAKEUP: '/img/icon-makeup.svg',
-    INDIVIDUAL_EDITING: '/img/icon-photo-edit.svg',
-    SUIT_RENTAL_FREE: '/img/icon-suit.svg',
-    ORIGINAL_FILES: '/img/icon-original-file.svg',
-    PARKING_AREA: '/img/icon-park.svg',
-  };
+  console.log(data);
 
   return (
-    <>
+    <div
+      css={css`
+        box-shadow: inset 0 0 10px red;
+      `}
+    >
       {/* 스튜디오 정보 */}
       <div css={StudioInfoTitleStyle}>
         <div>
@@ -68,7 +46,7 @@ const StudioInfo = ({ data, id }: IStudioInfo) => {
           <Bookmark id={Number(id)} count={data.bookmark_count} isBookmarked={false} />
         </div>
       </div>
-      <div css={StudioInfoStyle}>
+      <div css={StudioInfoStyle({ isPc })}>
         <dl>
           <div>
             <dt>
@@ -115,28 +93,8 @@ const StudioInfo = ({ data, id }: IStudioInfo) => {
       </div>
 
       {/* 홈 기본정보 - 매장 정보 */}
-      {isPc && (
-        <div css={optionsStyle}>
-          <p>매장 정보</p>
-          <div>
-            {data.options.length === 0
-              ? '수집중'
-              : data.options.map((optionItem) => (
-                  <Button
-                    key={optionItem}
-                    text={option[optionItem]}
-                    size="xsmall"
-                    width="fit"
-                    variant="white"
-                    iconSizeWidth="1.5rem"
-                    iconSizeHeight="1.5rem"
-                    icon={<img src={optionIcon[optionItem]} alt="매장정보" />}
-                  />
-                ))}
-          </div>
-        </div>
-      )}
-    </>
+      {isPc && <StudioOptions data={data} />}
+    </div>
   );
 };
 
@@ -170,7 +128,7 @@ const StudioInfoTitleStyle = css`
   }
 `;
 
-const StudioInfoStyle = css`
+const StudioInfoStyle = ({ isPc }: { isPc: boolean }) => css`
   position: relative;
 
   dl {
@@ -212,7 +170,7 @@ const StudioInfoStyle = css`
       }
     }
 
-    ${DividerStyle}
+    ${!isPc && DividerStyle}
   }
 `;
 
@@ -221,20 +179,4 @@ const SocialActionsStyle = css`
   gap: 2.4rem;
   ${TypoCapSmR}
   color: ${variables.colors.gray700};
-`;
-
-const optionsStyle = css`
-  padding: 2rem 0;
-  margin-bottom: 5rem;
-
-  & > p {
-    ${TypoTitleXsM};
-    margin-bottom: 1rem;
-  }
-
-  & > div {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 0.8rem;
-  }
 `;
